Fix for bug #2736

+++ sope-appserver/NGObjWeb/WebDAV/SaxDAVHandler.m
@@ -937,8 +937,8 @@ static BOOL heavyLog = NO;
 
 /* CDATA */
 
-- (void)characters:(unichar *)_chars length:(int)_len {
-  if (heavyLog) [self logWithFormat:@"got %i chars", _len];
+- (void)characters:(unichar *)_chars length:(NSUInteger)_len {
+  if (heavyLog) [self logWithFormat:@"got %"PRIuPTR" chars", _len];
   
   if (_len > 0 && (self->cdata != nil)) {
     NSString *s;
